get rid of quote report generator and just store in report form
